ORS-8: Reactored the the search bar to include a redux container component
