+++ src/components/search-bar/action.tsx
@@ -1,0 +1,6 @@
+export const fetchCategories = () => {
+  console.log('Hello World');
+  return {
+    type: 'apple',
+  }
+}
