+++ src/components/search-bar/component/styles.tsx
