+++ src/components/search-bar/component/types.tsx
@@ -19,6 +19,10 @@ export interface IOwnState {
   error: string
 }
 
+export interface IReduxProps {
+  'fetchCategories': () => void;
+}
+
 export interface IOptionType {
   value: string,
   label: string,
